WIP: dedicated Client Manager class
